refractor out the comment editor to a reusable component
