+++ lib/views/review-comment-view.js
@@ -1,25 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import {remote, shell} from 'electron';
-const {Menu, MenuItem} = remote;
-import {TextBuffer} from 'atom';
-
 
 import Timeago from './timeago';
 import Octicon from '../atom/octicon';
-import AtomTextEditor from '../atom/atom-text-editor';
 import GithubDotcomMarkdown from './github-dotcom-markdown';
 import EmojiReactionsController from '../controllers/emoji-reactions-controller';
 import {GHOST_USER} from '../helpers';
-import {addEvent} from '../reporter-proxy';
 import RefHolder from '../models/ref-holder';
+
+import EditableReview from './editable-review';
 
 export default class ReviewCommentView extends React.Component {
   static propTypes = {
     comment: PropTypes.object.isRequired,
     isPosting: PropTypes.bool.isRequired,
-    confirm: PropTypes.object.isRequired,
+    confirm: PropTypes.func.isRequired,
     tooltips: PropTypes.object.isRequired,
     renderEditedLink: PropTypes.func.isRequired,
     renderAuthorAssociation: PropTypes.func.isRequired,
@@ -38,44 +34,17 @@ export default class ReviewCommentView extends React.Component {
   }
 
   render() {
-    return this.state.editing ? this.renderEditor() : this.renderComment();
-  }
-
-  renderEditor() {
-    const buffer = new TextBuffer();
-    buffer.setText(this.props.comment.body);
-
     return (
-      <div className={cx('github-Review-editComment', {'github-Review-editComment--disabled': this.props.isPosting})}>
-        <AtomTextEditor
-          buffer={buffer}
-          lineNumberGutterVisible={false}
-          softWrapped={true}
-          autoHeight={true}
-          readOnly={this.props.isPosting}
-          refModel={this.refEditor}
-        />
-        <footer className="github-Review-editComment-footer">
-          <button
-            className="github-Review-editCommentCancelButton btn"
-            title="Cancel editing comment"
-            disabled={this.props.isPosting}
-            onClick={() => this.onCancel(buffer.getText() || '')}>
-            Cancel
-          </button>
-          <button
-            className="github-Review-updateCommentButton btn btn-primary"
-            title="Update comment"
-            disabled={this.props.isPosting}
-            onClick={() => this.onSubmitUpdate(buffer.getText() || '')}>
-            Update comment
-          </button>
-        </footer>
-      </div>
-    );
+      <EditableReview
+        originalContent={this.props.comment}
+        isPosting={this.props.isPosting}
+        confirm={this.props.confirm}
+        updateComment={this.props.updateComment}
+        render={this.renderComment}
+      />);
   }
 
-  renderComment() {
+  renderComment = showActionsMenu => {
     const comment = this.props.comment;
 
     if (comment.isMinimized) {
@@ -112,7 +81,7 @@ export default class ReviewCommentView extends React.Component {
           <Octicon
             icon="ellipses"
             className="github-Review-actionsMenu"
-            onClick={event => this.showActionsMenu(event, comment, author)}
+            onClick={event => showActionsMenu(event, comment, author)}
           />
         </header>
         <div className="github-Review-text">
@@ -131,74 +100,4 @@ export default class ReviewCommentView extends React.Component {
     );
   }
 
-  onCancel = text => {
-    if (text === this.props.comment.body) {
-      this.setState({editing: false});
-    } else {
-      const choice = this.props.confirm({
-        message: 'Are you sure you want to discard your unsaved changes?',
-        buttons: ['OK', 'Cancel'],
-      });
-      if (choice === 0) {
-        this.setState({editing: false});
-      }
-    }
-  }
-
-  onSubmitUpdate = text => {
-    if (text === this.props.comment.body || text === '') {
-      this.setState({editing: false});
-      return;
-    }
-    const didUpdateComment = () => this.setState({editing: false});
-    this.props.updateComment(this.props.comment.id, text, didUpdateComment);
-  }
-
-  showActionsMenu = (event, comment, author) => {
-    event.preventDefault();
-
-    const menu = new Menu();
-
-    menu.append(new MenuItem({
-      label: 'Edit',
-      click: () => this.setState({editing: true}),
-    }));
-
-    menu.append(new MenuItem({
-      label: 'Open on GitHub',
-      click: () => this.openOnGitHub(comment.url),
-    }));
-
-    menu.append(new MenuItem({
-      label: 'Report abuse',
-      click: () => this.reportAbuse(comment.url, author.login),
-    }));
-
-    menu.popup(remote.getCurrentWindow());
-  }
-
-
-  reportAbuse = (commentUrl, author) => {
-    return new Promise((resolve, reject) => {
-      const url = 'https://github.com/contact/report-content?report=' +
-        `${encodeURIComponent(author)}&content_url=${encodeURIComponent(commentUrl)}`;
-      shell.openExternal(url, {}, err => {
-        if (err) { reject(err); } else {
-          resolve();
-          addEvent('report-abuse', {package: 'github', component: this.constructor.name});
-        }
-      });
-    });
-  }
-
-  openOnGitHub = url => {
-    return new Promise((resolve, reject) => {
-      shell.openExternal(url, {}, err => {
-        if (err) { reject(err); } else {
-          resolve();
-          addEvent('open-comment-in-browser', {package: 'github', component: this.constructor.name});
-        }
-      });
-    });
-  }
 }
